feat: refactor module file
